Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -13,7 +13,7 @@ int main() {
     else if(c == '*')
         printf("%d\n", a * b);
     else if(c == '/')
-        printf("%f\n", a / b);
+        printf("%f\n", (float)a / b);
     else
         printf("error");
 }
